test: [step4] BodyDirectionStrategyTest 완료
